Finished

+++ List.h
@@ -1,3 +1,8 @@
+//Name: Luke Driscoll, Kai Itokazu
+//Student ID Number: 2344496, 2344742
+//Student Email: ldriscoll@chapman.edu, itokazu@chapman.edu
+//Class: CPSC-350-01 - Prof. German
+//Assignment 5: Registrars Office
 #ifndef LIST_H
 #define LIST_H
 
@@ -18,7 +23,7 @@ class List{
         virtual T remove(T key) = 0;
         virtual int find(T value) = 0;
         virtual bool isEmpty() = 0;
-        virtual size_t getSize() = 0;
+        virtual int getSize() = 0;
         //virtual void printList(bool printLink) = 0;
 };
